Update CollisionEvent.cc
Give secondary particles their random state *before* calling updateTrajectory.  Then we don't need to pass the original particle into updateTrajectory just to get access to the random state since we can use the random state of the particle we are updating.  This will change answers because we'll sample the random streams in a different order, but this pull request already changed answers when we changed to processing the original particle after the secondaries.